feat(#690): move performance_tuning.py from core/ to lib/ (frozen config dataclasses)
performance_tuning.py contains frozen performance config dataclasses
(ProfileTuning, EvictionTuning, etc.) with no kernel semantics.
Also fix 2 pre-existing mypy errors in service_wiring.py and agent_rpc_service.py.
