+++ src/nexus/services/agents/agent_rpc_service.py
@@ -338,8 +338,10 @@ class AgentRPCService:
     def _ensure_agent_registry(self) -> None:
         if self._agent_registry is not None:
             return
+        if self._session_factory is None:
+            raise RuntimeError("AgentRegistry not initialized and no session_factory available.")
         if self._record_store is None:
-            raise RuntimeError("AgentRegistry not initialized and no record_store available.")
+            raise RuntimeError("record_store is required to initialize AgentRegistry")
         from nexus.services.agents.agent_registry import AgentRegistry
 
         self._agent_registry = AgentRegistry(
