+++ tests/unit/services/test_eviction_manager.py
@@ -22,7 +22,7 @@ import pytest
 
 from nexus.contracts.agent_types import AgentRecord, AgentState, EvictionReason
 from nexus.contracts.qos import AgentQoS, QoSClass
-from nexus.core.performance_tuning import EvictionTuning
+from nexus.lib.performance_tuning import EvictionTuning
 from nexus.services.agents.eviction_manager import EvictionManager
 from nexus.services.agents.eviction_policy import LRUEvictionPolicy, QoSEvictionPolicy
 from nexus.services.agents.resource_monitor import PressureLevel
